+++ flowcraft/bin/report_POST.sh
@@ -27,14 +27,14 @@ then
     then
         json_str=$(cat $(pwd)/.report.json | sed 's/ //g' | sed s/\"/\'/g)
     else
-        json_str=$(cat $(pwd)/.report.json | sed 's/ /_/g' | sed s/\"/\'/g)
+        json_str=$(cat $(pwd)/.report.json | sed 's/ /%20/g' | sed s/\"/\'/g)
     fi
 fi
 
 # If a .versions file was populated, set the version_str variable
 if [ -s .versions ];
 then
-    version_str=$(< $(pwd)/.versions sed 's/ /_/g' | sed s/\"/\'/g)
+    version_str=$(< $(pwd)/.versions sed 's/ /%20/g' | sed s/\"/\'/g)
 fi
 
 if [ -s .command.trace ];
@@ -46,7 +46,7 @@ fi
 if [ ! "$json_str" = "{}" ] || [ ! "$version_str" = "[]" ] || [ ! "$trace_str" = "" ];
 then
     workdir=$(pwd)
-    json="{'project_id':'$projectid','pipeline_id':'$pipelineid','process_id':'$processid','sample_name':'$sample','report_json':$json_str,'current_user_name':'$username','current_user_id':'$userid','workdir':'$workdir','task':'$task','species':'$species','versions':$version_str,'trace':'$trace_str', 'overwrite': '$overwrite'}"
+    json="{'projectid':'$projectid','pipelineId':'$pipelineid','processId':'$processid','sample_name':'$sample','reportJson':$json_str,'username':'$username','userId':'$userid','workdir':'$workdir','task':'$task','processName':'$task','species':'$species','versions':$version_str,'trace':'$trace_str', 'overwrite': '$overwrite'}"
     echo \"${json}\" > .final.json
     {
         cat .final.json | curl -H  "Content-Type: application/json" -k -L -X POST -d @- $url > /dev/null