+++ flowcraft/flowcraft.py
@@ -15,6 +15,7 @@ try:
     from __init__ import __version__, __build__
     from generator.engine import NextflowGenerator, process_map
     from generator.inspect import NextflowInspector
+    from generator.report import FlowcraftReport
     from generator.recipe import brew_recipe, available_recipes
     from generator.pipeline_parser import parse_pipeline, SanityError
     from generator.process_details import proc_collector, colored_print
@@ -23,6 +24,7 @@ except ImportError:
     from flowcraft import __version__, __build__
     from flowcraft.generator.engine import NextflowGenerator, process_map
     from flowcraft.generator.inspect import NextflowInspector
+    from flowcraft.generator.report import FlowcraftReport
     from flowcraft.generator.recipe import brew_recipe, available_recipes
     from flowcraft.generator.pipeline_parser import parse_pipeline, \
         SanityError
@@ -130,6 +132,36 @@ def get_args(args=None):
         help="Pretty inspection mode that removes usual reporting processes."
     )
 
+    # REPORT MODE
+    reports_parser = subparsers.add_parser("report",
+                                           help="Broadcast the report of "
+                                                "a pipeline")
+    reports_parser.add_argument(
+        "-i", dest="report_file",
+        default="pipeline_report/pipeline_report.json",
+        help="Specify the path to the pipeline report JSON file."
+    )
+    reports_parser.add_argument(
+        "-u", "--url", dest="url", default="http://192.92.149.169:80/",
+        help="Specify the URL to where the data should be broadcast"
+    )
+    reports_parser.add_argument(
+        "--trace-file", dest="trace_file", default="pipeline_stats.txt",
+        help="Specify the nextflow trace file. Only applicable in combination "
+             "with --watch option."
+    )
+    reports_parser.add_argument(
+        "--log-file", dest="log_file", default=".nextflow.log",
+        help="Specify the nextflow log file. Only applicable in combination "
+             "with --watch option."
+    )
+    reports_parser.add_argument(
+        "-w", "--watch", dest="watch",  action="store_const", const=True,
+        help="Run the report in watch mode. This option will track the "
+             "generation of reports during the execution of the pipeline, "
+             "allowing for the visualization of the reports in real-time"
+    )
+
     if len(sys.argv) == 1:
         parser.print_help()
         sys.exit(1)
@@ -216,6 +248,9 @@ def copy_project(path):
 
     # Copy Helper scripts
     copy_tree(join(repo_dir, "lib"), join(target_dir, "lib"))
+
+    # Copy resources dir
+    copy_tree(join(repo_dir, "resources"), join(target_dir, "resources"))
 
     # Copy bin scripts
     copy_tree(join(repo_dir, "bin"), join(target_dir, "bin"))
@@ -322,6 +357,22 @@ def inspect(args):
         nf_inspect.broadcast_status()
 
 
+def report(args):
+
+    try:
+        fc_report = FlowcraftReport(
+            report_file=args.report_file,
+            trace_file=args.trace_file,
+            log_file=args.log_file,
+            watch=args.watch,
+            ip_addr=args.url)
+    except eh.ReportError as e:
+        logger.error(colored_print(e.value, "red_bold"))
+        sys.exit(1)
+
+    fc_report.broadcast_report()
+
+
 def main():
 
     args = get_args()
@@ -355,6 +406,9 @@ def main():
 
     if args.main_op == "inspect":
         inspect(args)
+
+    if args.main_op == "report":
+        report(args)
 
 
 if __name__ == '__main__':