+++ flowcraft/templates/assembly_report.py
@@ -54,7 +54,7 @@ def __get_version_pilon():
 
     try:
 
-        cli = ["java", "-jar", pilon_path , "--version"]
+        cli = ["java", "-jar", pilon_path, "--version"]
         p = subprocess.Popen(cli, stdout=PIPE, stderr=PIPE)
         stdout, _ = p.communicate()
 
@@ -310,24 +310,9 @@ class Assembly:
             contig_id = self._get_contig_id(contig)
             self.contig_boundaries[contig_id] = [c, c + len(seq)]
             c += len(seq)
-            xbars.append(
-                {
-                    "contig": contig_id,
-                    "position": c / window,
-                    "absPosition": c,
-                    "window": window
-                }
-            )
+            xbars.append((contig_id, c, contig))
 
-        # Get label contig for each window
-        labels = []
-        for i in range(0, self.summary_info["total_len"], window):
-            for contig, rg in self.contig_boundaries.items():
-                if rg[0] <= i < rg[1]:
-                    labels.append("{}_{}".format(contig, i))
-                    break
-
-        return labels, xbars
+        return xbars
 
     @staticmethod
     def _gc_prop(s, length):
@@ -348,7 +333,7 @@ class Assembly:
 
         return gc / length
 
-    def get_gc_sliding(self, window=500):
+    def get_gc_sliding(self, window=2000):
         """Calculates a sliding window of the GC content for the assembly
 
 
@@ -357,17 +342,9 @@ class Assembly:
         gc_res : list
             List of GC proportion floats for each data point in the sliding
             window
-        labels: list
-            List of labels for each data point
-        xbars : list
-            List of the ending position of each contig in the genome
-
         """
 
         gc_res = []
-
-        # Get contigID for each window position
-        labels, xbars = self._get_window_labels(window)
 
         # Get complete sequence to calculate sliding window values
         complete_seq = "".join(self.contigs.values()).lower()
@@ -377,9 +354,9 @@ class Assembly:
             seq_window = complete_seq[i:i + window]
 
             # Get GC proportion
-            gc_res.append(self._gc_prop(seq_window, len(seq_window)))
+            gc_res.append(round(self._gc_prop(seq_window, len(seq_window)), 2))
 
-        return gc_res, labels, xbars
+        return gc_res
 
     def _get_coverage_from_file(self, coverage_file):
         """
@@ -408,7 +385,7 @@ class Assembly:
                 else:
                     self.contig_coverage[header].append(coverage)
 
-    def get_coverage_sliding(self, coverage_file, window=500):
+    def get_coverage_sliding(self, coverage_file, window=2000):
         """
 
         Parameters
@@ -427,9 +404,6 @@ class Assembly:
         if not self.contig_coverage:
             self._get_coverage_from_file(coverage_file)
 
-        # Get contigID for each window position
-        labels, xbars = self._get_window_labels(window)
-
         # Stores the coverage results
         cov_res = []
 
@@ -440,9 +414,9 @@ class Assembly:
             # Get coverage values for current window
             cov_window = complete_cov[i:i + window]
             # Get mean coverage
-            cov_res.append(sum(cov_window) / len(cov_window))
+            cov_res.append(int(sum(cov_window) / len(cov_window)))
 
-        return cov_res, labels, xbars
+        return cov_res
 
 
 @MainWrapper
@@ -489,21 +463,26 @@ def main(sample_id, assembly_file, coverage_bp_file=None):
 
     if coverage_bp_file:
         try:
-            gc_sliding_data, gc_label, gc_xbars = assembly_obj.get_gc_sliding()
-            cov_sliding_data, cov_label, cov_xbars = \
-                assembly_obj.get_coverage_sliding(coverage_bp_file)
+            window = 2000
+            gc_sliding_data = assembly_obj.get_gc_sliding(window=window)
+            cov_sliding_data = \
+                assembly_obj.get_coverage_sliding(coverage_bp_file,
+                                                  window=window)
 
             # Get total basepairs based on the individual coverage of each
-            # contig bp
+            # contig bpx
             total_bp = sum(
                 [sum(x) for x in assembly_obj.contig_coverage.values()]
             )
 
             # Add data to json report
-            json_dic["plotData"][0]["data"]["gcSliding"] = \
-                [gc_sliding_data, gc_label, gc_xbars]
-            json_dic["plotData"][0]["data"]["covSliding"] = \
-                [cov_sliding_data, cov_label, cov_xbars]
+            json_dic["plotData"][0]["data"]["genomeSliding"] = {
+                "gcData": gc_sliding_data,
+                "covData": cov_sliding_data,
+                "window": window,
+                "xbars": assembly_obj._get_window_labels(window),
+                "assemblyFile": os.path.basename(assembly_file)
+            }
             json_dic["plotData"][0]["data"]["sparkline"] = total_bp
 
         except: