+++ flowcraft/templates/mashscreen2json.py
@@ -67,11 +67,11 @@ def main(mash_output, sample_id):
     for line in read_mash_output:
         tab_split = line.split("\t")
         identity = tab_split[0]
-        #shared_hashes = tab_split[1]
+        # shared_hashes = tab_split[1]
         median_multiplicity = tab_split[2]
-        #p_value = tab_split[3]
+        # p_value = tab_split[3]
         query_id = tab_split[4]
-        #query-comment should not exist here and it is irrelevant
+        # query-comment should not exist here and it is irrelevant
 
         # here identity is what in fact interests to report to json but
         # median_multiplicity also is important since it gives an rough
@@ -109,13 +109,20 @@ def main(mash_output, sample_id):
     output_json.close()
 
     json_dic = {
-        "sample_id": sample_id,
-        "patlas_mashscreen": filtered_dic
-        # TODO add information for report webapp
+        "tableRow": [{
+            "sample": sample_id,
+            "data": [{
+                "header": "Mash Screen",
+                "table": "plasmids",
+                "patlas_mashscreen": filtered_dic,
+                "value": len(filtered_dic)
+            }]
+        }],
     }
 
     with open(".report.json", "w") as json_report:
         json_report.write(json.dumps(json_dic, separators=(",", ":")))
+
 
 if __name__ == "__main__":